+++ build.sh
@@ -13,6 +13,7 @@ function usage()
 
 buildtype=release
 disable_pcapng=false
+enable_asan=false
 
 if [ -n "$ST_BUILD_DISABLE_PCAPNG" ];  then
     if [ $ST_BUILD_DISABLE_PCAPNG == "true" ]; then
@@ -21,10 +22,19 @@ if [ -n "$ST_BUILD_DISABLE_PCAPNG" ];  then
     fi
 fi
 
+if [ -n "$ST_BUILD_ENABLE_ASAN" ];  then
+    if [ $ST_BUILD_ENABLE_ASAN == "true" ]; then
+        enable_asan=true
+        buildtype=debug # use debug build as default for asan
+        echo "Enable asan check."
+    fi
+fi
+
 if [ -n "$1" ];  then
     case $1 in
       "debug")
            buildtype=debug
+           enable_asan=true
            ;;
       "debugoptimized")
            buildtype=debugoptimized
@@ -48,7 +58,7 @@ TEST_BUILD_DIR=${WORKSPACE}/build/tests
 PLUGINS_BUILD_DIR=${WORKSPACE}/build/plugins
 
 # build lib
-meson ${LIB_BUILD_DIR} -Dbuildtype=$buildtype -Ddisable_pcapng=$disable_pcapng
+meson ${LIB_BUILD_DIR} -Dbuildtype=$buildtype -Ddisable_pcapng=$disable_pcapng -Denable_asan=$enable_asan
 pushd ${LIB_BUILD_DIR}
 ninja
 sudo ninja install
@@ -56,7 +66,7 @@ popd
 
 # build app
 pushd app/
-meson ${APP_BUILD_DIR} -Dbuildtype=$buildtype
+meson ${APP_BUILD_DIR} -Dbuildtype=$buildtype -Denable_asan=$enable_asan
 popd
 pushd ${APP_BUILD_DIR}
 ninja
@@ -64,7 +74,7 @@ popd
 
 # build tests
 pushd tests/
-meson ${TEST_BUILD_DIR} -Dbuildtype=$buildtype
+meson ${TEST_BUILD_DIR} -Dbuildtype=$buildtype -Denable_asan=$enable_asan
 popd
 pushd ${TEST_BUILD_DIR}
 ninja
@@ -72,7 +82,7 @@ popd
 
 # build plugins
 pushd plugins/
-meson ${PLUGINS_BUILD_DIR} -Dbuildtype=$buildtype
+meson ${PLUGINS_BUILD_DIR} -Dbuildtype=$buildtype -Denable_asan=$enable_asan
 popd
 pushd ${PLUGINS_BUILD_DIR}
 ninja